chore moved the log.js source to utilities.
