+++ src/adapters/Cornerstone/Segmentation_4X.js
@@ -1,4 +1,4 @@
-import log from "../../log.js";
+import log from "../../utilities/log.js";
 import ndarray from "ndarray";
 import { BitArray } from "../../bitArray.js";
 import { datasetToBlob } from "../../datasetToBlob.js";
