+++ src/utilities/compression/rleSingleSamplePerPixel.js
@@ -1,4 +1,4 @@
-import log from "../../log.js";
+import log from "../log.js";
 
 /**
  * Encodes a non-bitpacked frame which has one sample per pixel.
