+++ test/integration/DicomMessage.readFile.test.js
@@ -1,7 +1,7 @@
 import fs from "fs";
 import path from "path";
 import dcmjs from "../../src/index.js";
-import { validationLog } from "./../../src/log.js";
+import { validationLog } from "../../src/utilities/log.js";
 
 // Ignore validation errors
 validationLog.setLevel(5);
